+++ solstice/src/main/java/dev/equo/ide/Catalog.java
@@ -13,8 +13,6 @@
  *******************************************************************************/
 package dev.equo.ide;
 
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import javax.annotation.Nullable;
 
@@ -25,28 +23,65 @@ public class Catalog implements Comparable<Catalog> {
 					"platform",
 					"https://download.eclipse.org/eclipse/updates/" + V,
 					"4.26",
-					"org.eclipse.platform.ide.categoryIU");
+					List.of("org.eclipse.platform.ide.categoryIU"));
 	public static final Catalog JDT =
-			new Catalog("jdt", PLATFORM, "org.eclipse.releng.java.languages.categoryIU", PLATFORM);
+			new Catalog(
+					"jdt", PLATFORM, List.of("org.eclipse.releng.java.languages.categoryIU"), PLATFORM);
 	public static final Catalog GRADLE_BUILDSHIP =
 			new Catalog(
 					"gradleBuildship",
 					"https://download.eclipse.org/buildship/updates/e423/releases/3.x/" + V,
 					"3.1.6.v20220511-1359",
-					"org.eclipse.buildship.feature.group",
+					List.of("org.eclipse.buildship.feature.group"),
 					JDT);
 	public static final Catalog PDE =
-			new Catalog("pde", PLATFORM, "org.eclipse.releng.pde.categoryIU", JDT);
+			new Catalog("pde", PLATFORM, List.of("org.eclipse.releng.pde.categoryIU"), JDT);
 
 	public static final Catalog KOTLIN =
 			new Catalog(
 					"kotlin",
 					"https://files.pkg.jetbrains.space/kotlin/p/kotlin-eclipse/main/" + V,
 					"0.8.21", // 0.8.24 is broken in many ways
-					Arrays.asList(
+					List.of(
 							"org.jetbrains.kotlin.feature.feature.group",
 							"org.jetbrains.kotlin.gradle.feature.feature.group"),
 					JDT);
+
+	public static final Catalog TM_TERMINAL =
+			new Catalog(
+					"tmTerminal",
+					"https://download.eclipse.org/tools/cdt/releases/11.0/cdt-" + V,
+					"11.0.0",
+					List.of(
+							"org.eclipse.tm.terminal.feature.feature.group",
+							"org.eclipse.tm.terminal.view.feature.feature.group"),
+					PLATFORM);
+
+	public static final Catalog CDT =
+			new Catalog(
+					"cdt",
+					TM_TERMINAL,
+					List.of(
+							"org.eclipse.launchbar.feature.group",
+							"org.eclipse.cdt.visualizer.feature.group",
+							"org.eclipse.cdt.unittest.feature.feature.group",
+							"org.eclipse.cdt.testsrunner.feature.feature.group",
+							"org.eclipse.cdt.native.feature.group",
+							"org.eclipse.cdt.msw.feature.group",
+							"org.eclipse.cdt.managedbuilder.llvm.feature.group",
+							"org.eclipse.cdt.llvm.dsf.lldb.feature.group",
+							"org.eclipse.cdt.launch.serial.feature.feature.group",
+							"org.eclipse.cdt.launch.remote.feature.group",
+							"org.eclipse.cdt.gnu.multicorevisualizer.feature.group",
+							"org.eclipse.cdt.feature.group",
+							"org.eclipse.cdt.debug.ui.memory.feature.group",
+							"org.eclipse.cdt.debug.standalone.feature.group",
+							"org.eclipse.cdt.debug.gdbjtag.feature.group",
+							"org.eclipse.cdt.core.autotools.feature.group",
+							"org.eclipse.cdt.cmake.feature.group",
+							"org.eclipse.cdt.build.crossgcc.feature.group",
+							"org.eclipse.cdt.autotools.feature.group"),
+					TM_TERMINAL);
 
 	private final String name;
 	private final String p2urlTemplate;
@@ -54,17 +89,8 @@ public class Catalog implements Comparable<Catalog> {
 	private final List<String> toInstall;
 	private final List<Catalog> requires;
 
-	Catalog(String name, Catalog copyFrom, String toInstall, Catalog... requires) {
+	Catalog(String name, Catalog copyFrom, List<String> toInstall, Catalog... requires) {
 		this(name, copyFrom.p2urlTemplate, copyFrom.latestVersion, toInstall, requires);
-	}
-
-	Catalog(
-			String name,
-			String p2urlTemplate,
-			String latestVersion,
-			String toInstall,
-			Catalog... requires) {
-		this(name, p2urlTemplate, latestVersion, Collections.singletonList(toInstall), requires);
 	}
 
 	Catalog(
@@ -81,7 +107,7 @@ public class Catalog implements Comparable<Catalog> {
 		}
 		this.latestVersion = latestVersion;
 		this.toInstall = toInstall;
-		this.requires = Arrays.asList(requires);
+		this.requires = List.of(requires);
 	}
 
 	public String getP2UrlTemplate() {